--- a/scripts/msys2-install-dependencies.sh
+++ b/scripts/msys2-install-dependencies.sh
@@ -44,6 +44,7 @@ pacboy --noconfirm --sync --needed \
     imagemagick:p \
     qt5-svg:p \
     qt5-multimedia:p \
+    tbb:p \
     python:p \
     python-pip:p