--- a/src/Feature.h
+++ b/src/Feature.h
@@ -21,6 +21,7 @@ public:
   static const Feature ExperimentalFastCsgExactCorefinementCallback;
   static const Feature ExperimentalFastCsgRemesh;
   static const Feature ExperimentalFastCsgRemeshPredictibly;
+  static const Feature ExperimentalManifold;
   static const Feature ExperimentalRoof;
   static const Feature ExperimentalInputDriverDBus;
   static const Feature ExperimentalLazyUnion;