--- a/src/geometry/GeometryUtils.cc
+++ b/src/geometry/GeometryUtils.cc
@@ -473,3 +473,28 @@ int GeometryUtils::findUnconnectedEdges(const std::vector<IndexedTriangle>& tria
 
   return edges.size();
 }
+
+Transform3d GeometryUtils::getResizeTransform(const BoundingBox &bbox, const Vector3d& newsize, const Eigen::Matrix<bool, 3, 1>& autosize)
+{
+  // Find largest dimension
+  int maxdim = 0;
+  for (int i = 1; i < 3; ++i) if (newsize[i] > newsize[maxdim]) maxdim = i;
+
+  // Default scale (scale with 1 if the new size is 0)
+  Vector3d scale(1, 1, 1);
+  for (int i = 0; i < 3; ++i) if (newsize[i] > 0) scale[i] = newsize[i] / bbox.sizes()[i];
+
+  // Autoscale where applicable
+  double autoscale = scale[maxdim];
+  Vector3d newscale;
+  for (int i = 0; i < 3; ++i) newscale[i] = !autosize[i] || (newsize[i] > 0) ? scale[i] : autoscale;
+
+  Transform3d t;
+  t.matrix() <<
+    newscale[0], 0, 0, 0,
+    0, newscale[1], 0, 0,
+    0, 0, newscale[2], 0,
+    0, 0, 0, 1;
+
+  return t;
+}