--- a/src/geometry/cgal/cgalutils-applyops.cc
+++ b/src/geometry/cgal/cgalutils-applyops.cc
@@ -10,6 +10,9 @@
 #include "printutils.h"
 #include "progress.h"
 #include "CGALHybridPolyhedron.h"
+#ifdef ENABLE_MANIFOLD
+#include "ManifoldGeometry.h"
+#endif
 #include "node.h"
 
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
@@ -191,6 +194,14 @@ bool applyHull(const Geometry::Geometries& children, PolySet& result)
           addPoint(vector_convert<K::Point_3>(p));
           return false;
         });
+#ifdef ENABLE_MANIFOLD
+    } else if (auto mani = dynamic_pointer_cast<const ManifoldGeometry>(chgeom)) {
+      points.reserve(points.size() + mani->numVertices());
+      mani->foreachVertexUntilTrue([&](auto& p) {
+          addPoint(vector_convert<K::Point_3>(p));
+          return false;
+        });
+#endif
     } else {
       const auto *ps = dynamic_cast<const PolySet *>(chgeom.get());
       if (ps) {
@@ -252,8 +263,9 @@ shared_ptr<const Geometry> applyMinkowski(const Geometry::Geometries& children)
 
         auto ps = dynamic_pointer_cast<const PolySet>(operands[i]);
         auto nef = dynamic_pointer_cast<const CGAL_Nef_polyhedron>(operands[i]);
-        if (auto hybrid = dynamic_pointer_cast<const CGALHybridPolyhedron>(operands[i])) {
-          nef = CGALUtils::createNefPolyhedronFromHybrid(*hybrid);
+
+        if (!nef) {
+          nef = CGALUtils::getNefPolyhedronFromGeometry(operands[i]);
         }
 
         if (ps) CGALUtils::createPolyhedronFromPolySet(*ps, poly);