--- a/src/geometry/cgal/cgalutils-applyops.cc
+++ b/src/geometry/cgal/cgalutils-applyops.cc
@@ -12,6 +12,7 @@
 #include "CGALHybridPolyhedron.h"
 #ifdef ENABLE_MANIFOLD
 #include "ManifoldGeometry.h"
+#include "manifoldutils.h"
 #endif
 #include "node.h"
 
@@ -241,6 +242,11 @@ bool applyHull(const Geometry::Geometries& children, PolySet& result)
  */
 shared_ptr<const Geometry> applyMinkowski(const Geometry::Geometries& children)
 {
+#if ENABLE_MANIFOLD
+  if (Feature::ExperimentalManifold.is_enabled()) {
+    return ManifoldUtils::applyMinkowskiManifold(children);
+  }
+#endif
   if (Feature::ExperimentalFastCsg.is_enabled()) {
     return applyMinkowskiHybrid(children);
   }