--- a/src/geometry/cgal/cgalutils-kernel.cc
+++ b/src/geometry/cgal/cgalutils-kernel.cc
@@ -27,6 +27,13 @@ CGAL_HybridKernel3::FT KernelConverter<CGAL::Epick, CGAL_HybridKernel3>::operato
 }
 
 template <>
+double KernelConverter<CGAL::Epick, CGAL_DoubleKernel>::operator()(
+  const double& n) const
+{
+  return n;
+}
+
+template <>
 CGAL::Epick::FT KernelConverter<CGAL_HybridKernel3, CGAL::Epick>::operator()(
   const CGAL_HybridKernel3::FT& n) const
 {