--- a/src/geometry/cgal/cgalutils.h
+++ b/src/geometry/cgal/cgalutils.h
@@ -30,6 +30,7 @@ Result vector_convert(V const& v) {
 
 namespace CGALUtils {
 
+CGAL_Nef_polyhedron *createNefPolyhedronFromPolySet(const PolySet& ps);
 bool applyHull(const Geometry::Geometries& children, PolySet& P);
 template <typename K>
 bool is_weakly_convex(const CGAL::Polyhedron_3<K>& p);
@@ -106,6 +107,7 @@ getCartesianConverter()
     FromKernel, ToKernel, KernelConverter<FromKernel, ToKernel>>();
 }
 shared_ptr<CGAL_Nef_polyhedron> createNefPolyhedronFromHybrid(const CGALHybridPolyhedron& hybrid);
+std::shared_ptr<CGALHybridPolyhedron> createHybridPolyhedronFromPolySet(const PolySet& ps);
 std::shared_ptr<CGALHybridPolyhedron> createMutableHybridPolyhedronFromGeometry(const std::shared_ptr<const Geometry>& geom);
 std::shared_ptr<const CGALHybridPolyhedron> getHybridPolyhedronFromGeometry(const std::shared_ptr<const Geometry>& geom);
 template <typename K>