… .circleci/config.yml
… .github/workflows/codeql-analysis.yml
… .github/workflows/linux.yml
… .github/workflows/macos-release.yml
… .github/workflows/macos-tests.yml
… README.md
… src/geometry/manifold/ManifoldGeometry.h
… src/geometry/manifold/manifold-applyops-minkowski.cc
… src/geometry/manifold/manifold-applyops.cc
… src/geometry/manifold/manifoldutils.cc
… src/geometry/manifold/manifoldutils.h
… src/glview/cgal/CGALRenderer.cc
… src/gui/MainWindow.cc
… src/io/export_3mf.cc
… src/io/export_stl.cc
… src/utils/parallel.h
… tests/CMakeLists.txt
… tests/regression/manifold-cgalpng/3mf-export-expected.png
… tests/regression/manifold-cgalpng/CSG-expected.png
… tests/regression/manifold-cgalpng/CSG-modules-expected.png
… tests/regression/manifold-cgalpng/difference-tests-expected.png
… tests/regression/manifold-cgalpng/example001-expected.png
… tests/regression/manifold-cgalpng/example002-expected.png
… tests/regression/manifold-cgalpng/example003-expected.png
… tests/regression/manifold-cgalpng/example004-expected.png
… tests/regression/manifold-cgalpng/example005-expected.png
… tests/regression/manifold-cgalpng/example006-expected.png
… tests/regression/manifold-cgalpng/example012-expected.png
… tests/regression/manifold-cgalpng/example016-expected.png
… tests/regression/manifold-cgalpng/example024-expected.png