This is introducing experimental support for the https://github.com/elalish/manifold library by @elalish.

**TL;DR**: It's [wicked fast](https://gist.github.com/ochafik/70a6b15e982b7ccd5a79ff9afd99dbcf#results), and ready to test / review. Maybe we get to ditch CGAL for most of the 3D rendering soon (except hulling & convex decomposition for minkowski), with 5-30x speedups over [fast-csg](https://github.com/openscad/openscad/pull/4087) (YMMV, multithreading works better on some models than others) and only _theoretical_ precision downgrade (TBC, please help test!)

You can download Windows / Linux binaries from the [last green CircleCI run](https://app.circleci.com/pipelines/github/openscad/openscad?branch=pull%2F4533) (e.g. [a Linux AppImage](https://output.circle-artifacts.com/output/job/48f63e84-44ad-4aa5-9d8d-aae50d0e23d9/artifacts/0/64-bit/OpenSCAD-2023.03.18.ai13977_PR4533-x86_64.AppImage), [a Win64 build](https://app.circleci.com/pipelines/github/openscad/openscad/3408/workflows/7aaea81c-6b68-4e16-b292-9502504e6f41/jobs/13978/artifacts), [a Win32 build](https://app.circleci.com/pipelines/github/openscad/openscad/3408/workflows/7aaea81c-6b68-4e16-b292-9502504e6f41/jobs/13976/artifacts)).

Note: fenced by the `manifold` experimental feature (needs enabling in the UI's settings or passing `--enable=manifold` in CLI`

### TODO:

- [x] Basic build & rendering plumbing
- [x] Find way for Manifold's multithreading to kick in ~~some confusion about TBB flags~~. This incidentally fixes https://github.com/openscad/openscad/issues/391, although more work could be needed to maximize core utilization.
- [x] Fix segfault on `tests/data/scad/issues/issue2342.scad` (sent fix to Manifold: https://github.com/elalish/manifold/pull/368 )
- [x] Fix Minkowski. **It's so fast now it's ridiculous.** (read below)
- [x] Implement Hull
- [x] Ensure Manifold uses batch operations. More specifically, let it build deeper internal trees by not forcing them to leaf nodes with our facet-based priority queue logic (it seems to have the same, along with [fast-union](https://github.com/openscad/openscad/pull/3636)-like logic to compose non intersecting operands).
- [x] Fix tests (enabling manifold for all but a some known issues)
- [x] Ensure builds on all platforms
- [x] Clean up code & submit for review
- [x] Run Benchmarks ([some results here](https://gist.github.com/ochafik/70a6b15e982b7ccd5a79ff9afd99dbcf#results), more to come)
- [ ] Fix mirror tests (prolly need to ask Manifold to flip faces if the matrix transform determinant is < 0 as we do with our other engines)
- [ ] Ensure works in WASM (CircleCI build env not updated yet)

### A note on precision

Manifold uses single-precision floats, which on paper is a far cry from OpenSCAD's numeric handling, which is double precision at worst, and as a general-rule uses exact rationals ([GMP](https://gmplib.org/) / [MPFR](https://www.mpfr.org/) + [CGAL](https://www.cgal.org/)) when doing CSG operations.

However, there's already substantial exceptions to OpenSCAD's exact numeric handling:
*   `minkowksi` and `hull` use double precision conversions and lose the exact rationals
*   `TransformNode`s (`multmatrix` / `translate` / `scale` / `rotate`) transform things in double then reconvert them to "exact" rationals. There's no reasonable way to get exact transforms in the general case anyway, e.g. to guarantee that the result of N rotations of `PI / N` around the origin or that applying two successive opposite transforms produces exactly the same solid (although that may appear to work in some cases).

That said, rounding errors from transforming simple precision coordinates could potentially snowball in trees that have many nested transforms.

The simple solution I'm rooting for here is to push the double-precision world transforms all the way down to pseudo-leaf nodes (i.e. actual leaves but also shared subtrees, minkowski, render, 2D-3D transitions), similarly to what I did in https://github.com/openscad/openscad/pull/3637.

I'll explore that route separately in https://github.com/openscad/openscad/pull/4561 (after initially trying to push the transform o the fly in `GeometryEvaluator` as what `CSGTreeEvaluator` does with the state; found out this messes up with caching so going a more traditional tree transform route).

### Licensing: GPLv2+ + Apache 2 = GPLv3

Manifold is under Apache 2 license, which [can be included in GPLv3 but not GPLv2 projects](https://www.apache.org/licenses/GPL-compatibility.html).

Since OpenSCAD source is GPLv2+, this means one can only release an OpenSCAD + Manifold binary under GPLv3. The same was already happening w/ CGAL (see [this thread](https://github.com/openscad/openscad/issues/2059)) so this doesn't change anything.

### Building & running

To use Manifold as an experimental rendering engine, build locally (make sure to [install the general prerequisites](https://github.com/openscad/openscad#building-openscad)) and run with `--enable=manifold` (or enable the feature in the UI's preferences)

```bash
git clone --recurse-submodules https://github.com/ochafik/openscad --branch manifold1 openscad-manifold
cd openscad-manifold
( cd submodules/manifold && git apply thrust.diff )

( \
  mkdir -p build && \
  cd build && \
  cmake \
    -DCMAKE_BUILD_TYPE=Release \
    -DCMAKE_INTERPROCEDURAL_OPTIMIZATION=1 \
    -DEXPERIMENTAL=1 \
    .. && \
    time make -j12 \
)
```

```
alias openscad=./build/OpenSCAD.app/Contents/MacOS/OpenSCAD # On Mac

openscad examples/Basics/CSG.scad --render -o out.stl '-D$fn=300'
# 50s

openscad examples/Basics/CSG.scad --render -o out.stl '-D$fn=300' --enable=fast-csg
# 2.8s

openscad examples/Basics/CSG.scad --render -o out.stl '-D$fn=300' --enable=manifold
# 0.8s (!)
```

(Interestingly enough, the rendering of that model breaks down at like `$fn=1000`. Might get back to this later to see if it's a Manifold issue or if it's about how we build spheres)

### minkowski, 20x faster (YMMV)

I've ported the OpenSCAD minkowski logic for the Manifold case, with a few upgrades:
- Fixed slight inefficiency in the original logic (points collection and conversion was being redone over and over)
- Sprinkled a bit of TBB-based parallelism (wink wink https://github.com/openscad/openscad/issues/391 ) since this brings that dependency anyway and there's less thread-unsafe CGAL numerics floating around to worry about. Used to decompose two operands to convex parts in parallel, and to execute all the cross products of decomposed part pairs in parallel. Can disable w/ `OPENSCAD_NO_PARALLEL=1` env var to see the difference (doesn't disable Manifold's parallel operations tho).
- Switched to CGAL's surface mesh for hulling

Long story short: **minkowski is orders of magnitude faster in lots of cases.**

For instance, @revarbat's [BOSL2 offset3d docs](https://github.com/revarbat/BOSL2/wiki/mutators.scad#module-offset3d) essentially say the function is so slow they don't even bother generating a preview. Well, the following takes 4sec on my M1 mac (at roughly 2.5 cores utilization), instead of 4m31sec (or 1m34sec with fast-csg):

```js
include <BOSL2/std.scad>
$fn=$preview ? 15 : 30;
offset3d(1)
    bottom_half()
        difference() {
            cube(10, center=true);
            xscale(0.9) yscale(1.1)
                sphere(d=10);
        }
```

<img width="276" alt="image" src="https://user-images.githubusercontent.com/273860/225490314-d90b50bc-d6b3-46bf-ac13-c0b8e8fcf783.png">

(Ngl, I struggled to find cool minkowski examples - adding [some here](https://gist.github.com/ochafik/70a6b15e982b7ccd5a79ff9afd99dbcf) - which this PR might help make more practical and widespread)

**That said** it looks like the CGAL Nef-based convex component decomposition we use is quick to crash with random models. Might muster the courage to file a bug soon