@pca006132 2023-02-27
For multithreading, can you try building manifold with python and run `▶ time python bindings/python/examples/run_all.py`?
On my Linux machine I got
```
Took 0.2ms for union_failure
Took 26.2ms for bricks
Took 35.3ms for cube_with_dents
Took 1200.6ms for maze
python bindings/python/examples/run_all.py  7.93s user 0.83s system 610% cpu 1.436 total
```
which is using a lot of cores, so I think thrust is working fine for this, perhaps it is an OS/build flag/TBB version thing?