@ochafik 2023-02-27
@elalish Ah yes, managing 2.5 cores with another model (condensed matter from https://github.com/openscad/openscad/pull/3641#issuecomment-780282019) , guess it will depend on the model tree shape. Getting same utilisation figures as you with Manifold’s perfTest so flags are fine.

I’m working on separate PRs for tree transforms to alter said shape anyway, and on octree scene partitioning to max out parallelism.