@gsohler 2023-02-28
Hi Oliver, thank you for attention.
I compiled exactly as documented, lately, only make  -j16 argument skipped  i had to install tbb-devel.
i have  only manifold experimental feature enabled.
other primitives  show fine, just polyhedron does not.
background is to find out how manifold handles self instersection by moving the last pt in pt list.
here is my model. it displays fine with default openscad

points=[
		[0,0,0], [10,0,0],
		[10,10,0],[0,10,0],
        [1,0,1], [9,0,1],
        [9,10,1],[1,10,1],
        [0,0,10],[1,0,10],
        [1,10,10],[0,10,10],
        [9,0,10],[10,0,10],
        [10,10,10],[9,10,10],
        [8,5,5]
	];

faces=[
	[3,2,1,0],[4,5,6,7],
    [8,9,10,11],[12,13,14,15],
    [0,1,13,12,5,4,9,8],
    [2,3,11,10,7,6,15,14],
    [1,2,14,13], [0,3,11,8],
    [5,12,15,6],[4,16,9],
    [9,16,10], [10,16,7],
    [7,16,4]
    ];

union()
{    
//cube();
//sphere();
polyhedron(points=points, faces=faces);
}