@ochafik 2023-02-28
@gsohler ah great, thanks for sharing! If you inspect the console output (easier in command line) you’ll see the warning `ERROR: [manifold] PolySet -> Manifold conversion failed: NotManifold`. I’d assume manifold skips solids in that state. 

I’ll see if it’s worth hack some mesh repairing attempt when manifold chokes on inputs. CGAL has some imperfect helpers for that.