@mconsidine 2023-03-05
Hi,
Quick question (and apologies if I should ask this elsewhere): looks like you're working with manifold and the fast-linalg2 lines at the same time.  Does one incorporate the other at the moment?  Compiled the manifold branch fine on Linux Mint 21 (Ubuntu) just fine.  But then wondered if the linalg2 branch was incorporating that as well.  TIA.  Great work ...
mconsidine