@elalish 2023-03-06
@gsohler Not currently but I'd like to, see [here](https://github.com/elalish/manifold/issues/289). It'll be a different algorithm though, because the parallelized one can't do self-intersection removal.