@mconsidine 2023-03-14
Hi,
When running cmake with the flags identified in the script, I get the following warning:
`CMake Warning:
  Manually-specified variables were not used by the project:

    THRUST_HOST_SYSTEM
    THRUST_MULTICONFIG_ENABLE_SYSTEM_TBB`

Can anyone tell me if this is expected or does it suggest I'm missing something somewhere?
TIA,
mconsidine