@ochafik 2023-03-14
@mconsidine thanks for reporting back on that, I'd missed that warning. Turns out these macros are already defined by Manifold's cmake build files, I got confused with [Thrust's instructions](https://github.com/NVIDIA/thrust/blob/main/thrust/cmake/README.md) / cmake variables that don't apply to the way Manifold includes it.

Hopefully you should see some multicore utilization regardless, depending on the model.