@ochafik 2023-03-15
@mconsidine you can be more gentle on your machine by reducing the build parallelism. `make -j12` means up to 12 parallel  compilations. You can bring it down to 2, or just drop the `-j` param for a fully serial build.