@mconsidine 2023-03-15
I'll get some numbers to you in the next couple of days.  The model (which is a WIP) is a reconstruction from blueprints of an old driving clock for a telescope.  So I suspect most of the render resources are being used on calculation of threads.  Haven't added gears yet.

Re: fast-linalg2 work.  Is/was that specific to fastcsg, or are there yet-more speed up opportunities when paired with manifold?  Just curious if I can start paying attention to only the manifold version.

Re core usage, I noticed that it seemed to be spreading the load out evenly.  Ie if one was at 24% they pretty much all were.  Ditto for other values.  So I *suppose* there could be a model that would max all of them out.  Or, put differently, if my machine only had 4 cores, I might bury it.

I do have a Linux laptop that I could run the experiment on.  Compiling with all cores but 2 has been nice (ie make -j$(nproc --ignore=2) I think does the trick).  But it was only with the latest work that I noticed all the RAM and swap partition getting used up.  So I don't think I'll try compiling on the laptop :)